week4-initial-deployment/requirements.txt
anyio==3.6.1
certifi==2022.9.14
click==8.1.3
fastapi==0.85.0
h11==0.14.0
idna==3.4
pip==22.1.2
pydantic==1.10.2
setuptools==63.4.1
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.3
wheel==0.37.1
fastapi
pydantic
python-multipart
requests
tweepy==4.10.1
transformers
python-dotenv
scipy
pyyaml
pandas
torch